refactor: moved signals mod to own file
